+++ app/auth/signup/page.tsx
@@ -1,4 +1,4 @@
-import { SignUpPage } from '@/pages/auth/signUp/signUpPage'
+import { SignUpPage } from '@/pages/auth/signUp/ui/signUpPage'
 
 export default function SignUp() {
   return <SignUpPage />
