+++ src/pages/auth/signUp/lib/useSignUpPage.tsx
@@ -1,0 +1,28 @@
+// @flow
+import { useSignUpMutation } from '@/entities/auth/api'
+import { useTranslation } from '@/shared/config'
+import { OutputSchema } from '@/features/auth/signup/lib/useSignUpForm'
+import { getErrorMessage } from '@/shared/lib/hooks'
+
+export const UseSignUpPage = () => {
+  const [signUp, { originalArgs, isSuccess, error }] = useSignUpMutation()
+  const email = originalArgs?.email || ''
+
+  const {
+    t: {
+      pages: {
+        auth: {
+          signUpPage: { errorMessages, ...page },
+        },
+      },
+    },
+  } = useTranslation()
+
+  const onSubmit = (data: OutputSchema) => {
+    signUp(data)
+  }
+
+  const customError = getErrorMessage({ errorMessages, error })
+
+  return { originalArgs, isSuccess, email, onSubmit, page, customError }
+}
