+++ src/pages/auth/signUp/ui/signUpPage.module.scss
@@ -4,7 +4,6 @@
   align-items: center;
 
   max-width: 378px;
-  margin-top: 24px;
   padding: 24px 24px 30px;
 
   background-color: var(--color-dark-500);
